Enable default deletion protection for RDS resources
Set deletion protection to true in order to safeguard against accidental deletion of databases when running terraform in automated pipeline. This may occur if code is pushed or merged by mistake.

This change reduces risk by making the deletion of databases an explicit operation where the deletion protection either is disabled when creating the database or disabled manually after creation.